pyproject.toml
[tool.poetry]
name = "movie-ranking-app"
version = "0.1.0"
description = ""
authors = ["Jakub Wujec <kuba1302@gmail.com>"]
readme = "README.md"
packages = [{include = "movie_ranking_app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
pydantic = "^1.10.4"
loguru = "^0.6.0"
pandas = "^1.5.2"
uvicorn = "^0.20.0"
matplotlib = "^3.6.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
